fix(plugin-python): send and read bigint as a digit string
bigint mapped to a bare int, so model_dump(mode="json") put a JSON number on
the wire. A ContractKit server's bigint schema only accepts a string ("123",
or the TypeScript SDK's "123n"): ServerKit's body parser revives "123n", the
Zod preprocess converts any other string, and a number is rejected with
"expected bigint, received number". Reading a "123n" back failed against int.

A new _scalars.py module, emitted with the base client, defines
BigInt = Annotated[int, BeforeValidator(...), PlainSerializer(str, when_used=
"json")]. It reads a digit string, "123n" or a number, and writes a plain digit
string, which is what the Kotlin, Swift and C# SDKs write. Models import it,
and the client imports it wherever a body, query param or response type
carries a bigint. Relative imports now render after the stdlib ones.

No generated server can put a bigint in a response today: the Koa and Fastify
routers leave serialization to JSON.stringify, which throws on a BigInt. The
reader accepts every form so it keeps working whichever way that is fixed.

+++ packages/output-tests/tests/__snapshots__/python/pysdk/_models_billing.py
@@ -6,6 +6,7 @@ from uuid import UUID
 from typing import Literal
 from pydantic import BaseModel, ConfigDict, Field
 from decimal import Decimal
+from ._scalars import BigInt
 
 # A customer payment
 class Payment(BaseModel):
@@ -14,7 +15,7 @@ class Payment(BaseModel):
     id: UUID
     amount: float
     unit_price: Decimal = Field(alias="unitPrice")
-    quantity: int
+    quantity: BigInt
     created_at: datetime = Field(alias="createdAt")
     processing_time: timedelta | None = Field(alias="processingTime", default=None)
     status: Literal["pending", "completed", "failed"] | None = Field(default="pending")
@@ -24,7 +25,7 @@ class PaymentInput(BaseModel):
 
     amount: float
     unit_price: Decimal = Field(alias="unitPrice")
-    quantity: int
+    quantity: BigInt
     created_at: datetime = Field(alias="createdAt")
     processing_time: timedelta | None = Field(alias="processingTime", default=None)
     status: Literal["pending", "completed", "failed"] | None = Field(default="pending")
